feat: Move version variable assignment to verifyRelease
