+++ lib/verifyRelease.js
